requirements.txt
PyQt5>=5.15
openai>=1.0
ddgs>=1.0
dropbox>=11.0
PyGithub>=2.0
apscheduler>=3.10
requests>=2.31
